+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.2 (unreleased)
 
+- Added `encode` option to `Lockbox::Encryptor`
 - Added `Lockbox::Audit` model
 
 ## 0.3.1 (2019-12-26)